+++ src/components/SpeechBubble/SpeechBubbleDetail.tsx
