+++ src/components/SpeechBubble/SpeechBubbleThumbnail.tsx
